fix(flexlb): widen the round-robin cursor to long and de-vacuum its concurrency test
The batch/select cursor was an AtomicInteger; like FePool it now uses an
AtomicLong so it never wraps at any realistic QPS. The old int wrapped every
~2^32 picks and, since 2^32 is not a multiple of the pool size, produced a
one-off round-robin discontinuity at the wrap point. Math.floorMod keeps the
index non-negative regardless.

selectBatch_atomic_range_does_not_overlap_with_concurrent_selects was vacuous:
with pool==batch==4, four consecutive integers mod 4 always cover {0,1,2,3} for
any start, so its only check (batchWorkers.size()==4) held even under a
non-atomic cursor. Rewritten to use a pool larger than the total picks per
trial, so each consumed cursor value maps to a distinct worker; it now asserts
every pick (batch + concurrent singles) is distinct and the cursor advances by
exactly the total, which a non-atomic reservation fails. Mutation-verified: a
get()+addAndGet() cursor turns it red.

+++ rtp_llm/flexlb/flexlb-sync/src/main/java/org/flexlb/balance/strategy/RoundRobinLoadBalancer.java
@@ -16,7 +16,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * Cursor-based round-robin load balancer.
@@ -52,7 +52,14 @@ public class RoundRobinLoadBalancer implements BatchLoadBalancer {
     private final EngineWorkerStatus engineWorkerStatus;
     /** Engine type is fixed at boot; resolved once instead of a config lookup per selection. */
     private final boolean embeddingEngine;
-    private final Map<String, AtomicInteger> cursors = new ConcurrentHashMap<>();
+    /**
+     * Per-(role, group) rotation cursors. Each is a {@code long} (not {@code int}) so it never wraps
+     * at any realistic QPS (2^63 picks), mirroring {@link org.flexlb.dispatcher.FePool}: an {@code int}
+     * cursor wrapped every ~2^32 picks and, because 2^32 is not a multiple of the pool size, produced
+     * a one-off RR discontinuity at the wrap point. {@link Math#floorMod(long, int)} keeps the index
+     * non-negative regardless.
+     */
+    private final Map<String, AtomicLong> cursors = new ConcurrentHashMap<>();
 
     public RoundRobinLoadBalancer(EngineWorkerStatus engineWorkerStatus, ConfigService configService) {
         this.engineWorkerStatus = engineWorkerStatus;
@@ -60,9 +67,9 @@ public class RoundRobinLoadBalancer implements BatchLoadBalancer {
         LoadBalanceStrategyFactory.register(LoadBalanceStrategyEnum.ROUND_ROBIN, this);
     }
 
-    private AtomicInteger cursor(RoleType roleType, String group) {
+    private AtomicLong cursor(RoleType roleType, String group) {
         String key = group == null ? roleType.name() : roleType.name() + '|' + group;
-        return cursors.computeIfAbsent(key, k -> new AtomicInteger(0));
+        return cursors.computeIfAbsent(key, k -> new AtomicLong(0));
     }
 
     @Override
@@ -85,7 +92,7 @@ public class RoundRobinLoadBalancer implements BatchLoadBalancer {
             return new ArrayList<>();
         }
         int aliveSize = alive.size();
-        int start = cursor(roleType, group).getAndAdd(count);
+        long start = cursor(roleType, group).getAndAdd(count);
         List<BatchScheduleTarget> targets = new ArrayList<>(count);
         for (int i = 0; i < count; i++) {
             int idx = Math.floorMod(start + i, aliveSize);
